app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.15.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
black==25.1.0
    # via -r requirements.in
boto3==1.37.35
    # via -r requirements.in
botocore==1.37.35
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   opensearch-py
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
    #   uvicorn
cuid==0.4
    # via -r requirements.in
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.2.0
    # via pydantic
events==0.5
    # via opensearch-py
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via -r requirements.in
flake8==7.2.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   requests
inflection==0.5.1
    # via pyairtable
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2024.10.1
    # via jsonschema
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
newrelic==10.9.0
    # via -r requirements.in
numpy==2.2.4
    # via pandas
opensearch-py==2.8.0
    # via -r requirements.in
packaging==24.2
    # via
    #   black
    #   build
pandas==2.2.3
    # via -r requirements.in
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.7
    # via black
psycopg2==2.9.10
    # via -r requirements.in
pyairtable==3.1.1
    # via -r requirements.in
pycodestyle==2.13.0
    # via flake8
pydantic[email]==2.11.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pyairtable
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.8.1
    # via -r requirements.in
pyflakes==3.3.2
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   opensearch-py
    #   pandas
python-dotenv==1.1.0
    # via
    #   pydantic-settings
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   opensearch-py
    #   pyairtable
retrying==1.3.4
    # via -r requirements.in
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via
    #   python-dateutil
    #   retrying
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
tomli==2.2.1
    # via
    #   black
    #   build
    #   pip-tools
typing-extensions==4.13.2
    # via
    #   alembic
    #   anyio
    #   black
    #   fastapi
    #   pyairtable
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   botocore
    #   opensearch-py
    #   pyairtable
    #   requests
uvicorn[standard]==0.34.1
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
